requirements.txt
pyserial
pyside6
rich
